BUG FIX: Renamed to Multiple_Memberships() and allowing PHPStan testing without WP loaded
